Changed test command to default to koltin. For java web3j generate-test --java.
