+++ src/main/java/org/web3j/console/Runner.java
@@ -18,14 +18,14 @@ import org.web3j.codegen.TruffleJsonFunctionWrapperGenerator;
 import org.web3j.console.config.CliConfig;
 import org.web3j.console.project.ProjectCreator;
 import org.web3j.console.project.ProjectImporter;
-import org.web3j.console.project.java.JavaTestCreator;
+import org.web3j.console.project.UnitTestCreator;
 import org.web3j.console.update.Updater;
 import org.web3j.utils.Version;
 
 import static org.web3j.codegen.SolidityFunctionWrapperGenerator.COMMAND_SOLIDITY;
 import static org.web3j.console.project.ProjectCreator.COMMAND_NEW;
 import static org.web3j.console.project.ProjectImporter.COMMAND_IMPORT;
-import static org.web3j.console.project.java.JavaTestCreator.COMMAND_GENERATE_TESTS;
+import static org.web3j.console.project.UnitTestCreator.COMMAND_GENERATE_TESTS;
 import static org.web3j.utils.Collection.tail;
 
 /** Main entry point for running command line utilities. */
@@ -86,7 +86,7 @@ public class Runner {
                     ContractAuditor.main(tail(args));
                     break;
                 case COMMAND_GENERATE_TESTS:
-                    JavaTestCreator.main(args);
+                    UnitTestCreator.main(tail(args));
                     break;
                 default:
                     Console.exitError(USAGE);
