+++ src/test/java/org/web3j/console/project/java/JavaTestCreatorTest.java
@@ -25,6 +25,7 @@ import org.junit.jupiter.api.io.TempDir;
 import picocli.CommandLine;
 
 import org.web3j.console.project.ProjectImporter;
+import org.web3j.console.project.UnitTestCreator;
 import org.web3j.console.project.utils.ClassExecutor;
 
 import static java.io.File.separator;
@@ -74,10 +75,10 @@ public class JavaTestCreatorTest extends ClassExecutor {
                         .start()
                         .waitFor();
         Assertions.assertEquals(0, exitCode);
-        final String[] unitTestsArgs = {"-i=" + pathToJavaWrappers, "-o=" + tempDirPath};
+        final String[] unitTestsArgs = {"--java", "-i=" + pathToJavaWrappers, "-o=" + tempDirPath};
         int testsExitCode =
                 executeClassAsSubProcessAndReturnProcess(
-                                JavaTestCreator.class,
+                                UnitTestCreator.class,
                                 Collections.emptyList(),
                                 Arrays.asList(unitTestsArgs))
                         .inheritIO()
